Implement IR file queue, load last requested file
This handles a case when a user asks to load a IR, while
the plugin's worker is busy load a file. Previously the
request was ignored.

This also fixes the case for LV2 hosts that do not call run()
after initial state-restore (e.g. MOD-host), so the work-response
is delivered late.