Fix fast restart bug (can happen during state restore)
Convolver::reset() is issued from Plugin::activate()
during session load, directly after loading the state.

At this point in time the Convolver is configured with
thread running, yet the reset restarts threads.

Since the threads are created with PTHREAD_CREATE_DETACHED,
it was possible that old threads are still running, when
new ones are started, which can wreak havoc leading to
memory corruption.